% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakdowngraph_func.R
\name{breakdown}
\alias{breakdown}
\title{Title}
\usage{
breakdown(all.frag, plot = TRUE, start = 0, end = 1)
}
\arguments{
\item{all.frag}{}

\item{plot}{default is TRUE}

\item{start}{passed to rainbow() function}

\item{end}{passed to rainbow() function}
}
\value{
Just a graphing function
}
\description{
Title
}
