% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betterSimilarityCalc_func.R
\name{fragCountSim}
\alias{fragCountSim}
\title{Wrapper function}
\usage{
fragCountSim(spec.top, spec.bottom, y = 0, z = 1, t = 0.005,
  replace.na = 0, b = 0.05, min.frag = 0)
}
\arguments{
\item{spec.top}{MSMS spectrum 1}

\item{spec.bottom}{MSMSpectrum 2}

\item{y}{m/z weighting factor for simTableWeighting(). Default is 0}

\item{z}{intensity weighting factor for simTableWeighting(). Default is 1}

\item{t}{m/z tolerance for merging factors for simTable(). Default is 0.005}

\item{replace.na}{value to replace NAs with. Default is 0}

\item{b}{relative intensity cutoff for relCutoff(). Default is 0.05.}

\item{min.frag}{the minimum number of matched fragments required. If number of matched fragments is less than this value, 
score is automatically 0. Default is 0}
}
\value{
List with the the similarity score (sum of number of matched fragments) and data frame showing output from each step
 (i.e., merging and intensity cutoff). First column is m/z, then the top and bottom weighted values used for the similarity
 score calculation. Fourth and fifth columns show the result after applying the relative intensity cutoff, followed by
 if fragments are matching. Final two column show the original input intensities prior to any processing.
}
\description{
Calculates similarity score of two MSMS spectra using count of matched fragments
}
\examples{

}

