% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCutoff.R
\name{getCutoff}
\alias{getCutoff}
\title{Function to get the Similarity Score Threshold for desired FPR}
\usage{
getCutoff(fptp, minFPR = 0)
}
\arguments{
\item{fptp}{output from the PandRdat function. A list with each of the scenarios to evaluate. Will also accept a data frame}

\item{minFPR}{the desired minimum false positive rate}
}
\value{
a vector equal to the length of fptp input, with the minimum similarity score needed for each scenario to reach the desired false positive rate
}
\description{
Function to get the Similarity Score Threshold for desired FPR
}
\examples{

}
