% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergedCEdata_func.R
\name{mergedCEdata}
\alias{mergedCEdata}
\title{Merged CE data for comparison}
\usage{
mergedCEdata(all.frag, pairs, dmz = 0.001, intensity.type)
}
\arguments{
\item{all.frag}{}

\item{pairs}{}

\item{dmz}{}

\item{intensity.type}{}
}
\value{
list of two lists, first with merged spectra of parent, second with merged spectra of TP. New function mergedSpectra 
is better because doesn't require input of the pairs.
}
\description{
Merged CE data for comparison
}

