% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergedSpectra_func.R
\name{mergedSpectra}
\alias{mergedSpectra}
\title{Title}
\usage{
mergedSpectra(all.frag, intensity.type = "absolute", dmz = 0.001)
}
\arguments{
\item{all.frag}{}

\item{intensity.type}{}

\item{dmz}{}
}
\description{
Title
}
\examples{

}
