% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betterSimilarityCalc_func.R
\name{MSsim}
\alias{MSsim}
\title{Similarity score calculation}
\usage{
MSsim(alignment)
}
\arguments{
\item{alignment}{merged fragment table, for example output from simTable(), relCutoff(), or simTableWeighting(). 
The two vectors which are being compared should be in columns 2 and 3}
}
\value{
correlation coeffiecient of the input vectors
}
\description{
Similarity score calculation
}
\details{
correlation coeffiecient is calculated as (u %*% v) / ( sqrt(sum(u*u)) * sqrt(sum(v*v)) ) where u in the first input vector and
v is the second input vector and %*% is the dot product
}

