% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betterSimilarityCalc_func.R
\name{OrgMSSim}
\alias{OrgMSSim}
\title{Calculates similarity score of two MSMS spectra using modified cosine}
\usage{
OrgMSSim(spec.top, spec.bottom, y = 0, z = 1, t = 0.005, replace.na = 0,
  b = 0.05, min.frag = 0)
}
\arguments{
\item{spec.top}{MSMS spectrum 1}

\item{spec.bottom}{MSMSpectrum 2}

\item{y}{m/z weighting factor for simTableWeighting(). Default is 0}

\item{z}{intensity weighting factor for simTableWeighting(). Default is 1}

\item{t}{m/z tolerance for merging factors for simTable(). Default is 0.005}

\item{replace.na}{value to replace NAs with. Default is 0}

\item{b}{relative intensity cutoff for relCutoff(). Default is 0.05.}
}
\value{
List with the correlation coefficient (ie the similarity score) and data frame showing output from each step (ie merging, intensity
cutoff, and weighting)
}
\description{
Calculates similarity score of two MSMS spectra using modified cosine
}
\examples{

}

