% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PandR_func.R
\name{PandR}
\alias{PandR}
\title{PandR function for looking at FPR vs. TPR.}
\usage{
PandR(realdata, randdata, steps = 0.001, metricList = metricList,
  specScenario = "AllFragments", PandRplot = TRUE,
  legend.lab = metricList, save.data = TRUE, auc = TRUE)
}
\arguments{
\item{realdata}{}

\item{randdata}{}

\item{steps}{}

\item{metricList}{}

\item{specScenario}{}

\item{PandRplot}{}

\item{legend.lab}{}

\item{save.data}{}

\item{auc}{}
}
\value{
Large wrapper for all PandR functions. Has largely been replaced by ROC_func
}
\description{
PandR function for looking at FPR vs. TPR.
}
