% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PandRdat_func.R
\name{PandRdat}
\alias{PandRdat}
\title{Generation of PandR data for scenario evaluation}
\usage{
PandRdat(realdata, randdata, metricList = metricList, steps = 0.001,
  specScenario = "AllFragments", save.data = TRUE)
}
\arguments{
\item{realdata}{}

\item{randdata}{}

\item{metricList}{}

\item{steps}{}

\item{specScenario}{}

\item{save.data}{}
}
\description{
Generation of PandR data for scenario evaluation
}
