% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betterSimilarityCalc_func.R
\name{relCutoff}
\alias{relCutoff}
\title{Relative intensity cutoff}
\usage{
relCutoff(alignment, b = 0.05, replace.na = 0)
}
\arguments{
\item{alignment}{merged fragment table, for example output from simTable()}

\item{b}{relative intensity cutoff, default is 0.05. all intensities below this value will be replaced with 0.}
}
\value{
data frame with fragment m/z values, intensity in MSMS spectrum 1, and intensity in MSMS spectrum 2
}
\description{
Relative intensity cutoff
}
\examples{

}

