% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC_func.R
\name{ROC}
\alias{ROC}
\title{ROC function for looking at FPR vs. TPR and graphing possibilites}
\usage{
ROC(realdata, randdata, metricList = metricList, steps = 0.001,
  specScenario = "AllFragments", PandRplot = TRUE,
  legend.lab = metricList, save.data = TRUE, aucs = TRUE, fPvfN = TRUE)
}
\arguments{
\item{realdata}{}

\item{randdata}{}

\item{metricList}{}

\item{steps}{}

\item{specScenario}{}

\item{PandRplot}{}

\item{legend.lab}{}

\item{save.data}{}

\item{aucs}{}

\item{fPvfN}{}
}
\description{
ROC function for looking at FPR vs. TPR and graphing possibilites
}
