% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sameCEanalysis_func.R
\name{sameCEanalysis}
\alias{sameCEanalysis}
\title{Analysis of Single Collion Energies}
\usage{
sameCEanalysis(pairs, nowp, nowt, sameCE, wMD = FALSE, y = 1, z = 1,
  t = 0.005, b = 0.5, replace.na = 0, sink = FALSE)
}
\arguments{
\item{pairs}{Data frame containing the pairs which are to be compared. First column should be "pair_ID", second column "TP_ID",
and the third column "par_ID".}

\item{nowp}{List which contains all the MSMS spectra for the parent compounds. Output from sameCEdata function}

\item{nowt}{List which contains all the MSMS spectra for the TP compounds. Output from sameCEdata function}

\item{sameCE}{List which contains the measurement information for each of the collision energies. Output from sameCEdata function}

\item{wMD}{Should the spectra of the TP be shifted by the mass difference of the pair? Default is FALSE}

\item{y}{m/z weighting factor for simTableWeighting(). Default is 0}

\item{z}{intensity weighting factor for simTableWeighting(). Default is 1}

\item{t}{m/z tolerance for merging factors for simTable(). Default is 0.005}

\item{b}{relative intensity cutoff for relCutoff(). Default is 0.05.}

\item{replace.na}{value to replace NAs with. Default is 0}

\item{sink}{Should the data be output into a .txt file? Default in FALSE}
}
\value{
List with the first entry the summary data frame with scores of all pairs at each collision energy. Second entry is a list, with
an entry for each of the collision energies analyzed with the measurement information.
}
\description{
Analysis of Single Collion Energies
}
