% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sameCEdata_func.R
\name{sameCEdata}
\alias{sameCEdata}
\title{Producing the list for comparing MS/MS spectra from same CEs for the pairs. Largely replaced by singleCESpectra_func}
\usage{
sameCEdata(all.frag, pairs)
}
\arguments{
\item{all.frag}{}

\item{pairs}{}
}
\value{
Returns list of three lists. First is the summary of each collision energy, second is the data of parents at all 
CEs, third is the data of TPs at all CEs.
}
\description{
Producing the list for comparing MS/MS spectra from same CEs for the pairs. Largely replaced by singleCESpectra_func
}
\examples{

}
