% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betterSimilarityCalc_func.R
\name{simTable}
\alias{simTable}
\title{Builds the merged fragment table}
\usage{
simTable(spec.top, spec.bottom, t = 0.005, replace.na = 0)
}
\arguments{
\item{spec.top}{MSMS spectrum 1}

\item{spec.bottom}{MSMS spectrum 2}

\item{t}{m/z tolerance for merging fragments. Default is 0.005}

\item{replace.na}{value to replace NAs with. Default is 0}
}
\value{
data frame with fragment m/z values, intensity in MSMS spectrum 1, and intensity in MSMS spectrum 2
}
\description{
Builds the merged fragment table
}
\examples{

}

