% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betterSimilarityCalc_func.R
\name{simTableWeighting}
\alias{simTableWeighting}
\title{Weighting of m/z and intensity}
\usage{
simTableWeighting(alignment, y = 0, z = 1)
}
\arguments{
\item{alignment}{merged fragment table, for example output from simTable() or relCutoff()}

\item{y}{weighting factor for m/z. Default is 0}

\item{z}{weight factor for intensity. Default is 1}
}
\value{
data frame with fragment m/z values, weighted vector for MSMS spectrum 1, and weighted vector of MSMS spectrum 2
}
\description{
Weighting of m/z and intensity
}
\details{
In order to compare spectra in MSsim function, vectors need to be generated for each spectra.
These vectors can be generated from only the intensities (default; y = 0, z = 1) or using different weighting factors. For example,
the dot product similarity scoring from the NIST database uses y = 3, z = 0.6 and from MassBank y = 2, z = 0.5.
Here the input intensity values for each spectrum are replaced by the weighted vector. The weighted vector is calculated as:

(m/z)^y * (intensity)^z
}
\examples{

}
