% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleCESpectra_func.R
\name{singleCESpectra}
\alias{singleCESpectra}
\title{spectra from single CE function}
\usage{
singleCESpectra(all.frag, energy = "15", dmz = 0.001)
}
\arguments{
\item{all.frag}{}

\item{energy}{collision energy desired}

\item{dmz}{}
}
\description{
spectra from single CE function
}
\examples{

}
